Updated JQuery version to 3.3.1
Updated JQuery to 3.3.1 such that it is compatible with DataTables
plugin. List of map projects is now populated, searchable, sortable,
and properly paged.

Updated menu.js to remove calls to deprecated / removed JQuery functions
($.browser).  We may need to revisit this - IE 6/7 users are unlikely
to experience proper functionality of the dropdown menu.

Included all necessary files for DataTables to function properly.